feat(admin): setup api route for formula
